backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.9.0
httpx==0.27.0
anthropic==0.18.0
resend==0.8.0
python-dotenv==1.0.0
pydantic==2.5.3